* Moved helper functions to * Removed namespace of helpers functions and moved to the global namesapce. * Adjusted helper functions usage. * Introduced  and  new functions.
