+++ src/Infrastructure/Component/Theme/ThemeComponent.php
@@ -16,15 +16,13 @@ namespace Webify\Base\Infrastructure\Component\Theme;
 use Webify\Base\Domain\Service\Theme\ThemeServiceInterface;
 use yii\base\Theme as BaseTheme;
 
-use function Webify\Base\Infrastructure\set_alias;
-
 /**
  * Theme abstraction class based on framework theme class.
  */
 abstract class ThemeComponent extends BaseTheme implements ThemeServiceInterface
 {
 	/**
-	 * Initializes the theme component by setting the current theme path as `@Theme` alias.
+	 * Initialises the theme component by setting the current theme path as `@Theme` alias.
 	 */
 	public function init(): void
 	{
